Ran linter
