+++ src/pages/index.tsx
@@ -3,13 +3,14 @@ import { NextPage } from "next";
 import Footer from "src/components/Footer";
 import EventsContainer from "src/components/EventsContainer";
 import Header from "src/components/Header";
+import DummyDate from "src/components/DummyDate";
 
 const Home: NextPage = () => {
     return (
         <div className="container">
             <Header />
             <div style={{ height: 100 }}>This is the home page with the list of events</div>
-
+            <DummyDate />
             <EventsContainer />
             <Footer />
